refactor: extract shared diagnosis core; unblock the step loop from observer passes
Post-review cleanup of the observer feature: the vocabulary, diagnosis
shape, terminal ladder, prompt vars, and filed-task frame shared by
selfImprove and the observer move to diagnosisCore.js — a deliberate leaf
outside the session.js import cycle, so OBSERVER_AREAS derives from
SELF_IMPROVE_AREAS instead of hand-mirroring it. Mid-run observer passes
now ride a per-run promise chain (drained before every terminal frame)
instead of blocking the next pipeline step on a heavy-model call, budget
exhaustion latches further mid-run passes off instead of re-reading the
usage file each step, and series.js shares one bounded task-ref sanitizer
between the two markers.
